wip: flows and state
Good news! I think I'm understanding flows now. Bad news! Little to no progress, and I have to change many structures! Yay! The problems of being a dumbass.

Another good thing is that I feel I'm learning lots of new things! Lets hope I don't forget them in the near future!

+++ app/src/main/java/com/example/project_shelf/app/entity/Product.kt
@@ -1,10 +1,8 @@
 package com.example.project_shelf.app.entity
 
-import java.math.BigDecimal
-
 data class Product(
     val id: Long,
     val name: String,
-    val defaultPrice: BigDecimal,
+    val defaultPrice: Long,
     val stock: Int,
 )
